Add event in Swanley, UK
Please check thoroughly - I've copied the format from other events and am a complete amateur on this. 

This particular event reoccurs, with bookings on 

9th, 16th, 23rd & 30th March

All times and locations, and booking remain the same for the other dates in March.
so may be easier to do either one entry for all days, or just alter the date.